% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errum-package.R
\docType{package}
\name{errum-package}
\alias{errum-package}
\alias{_PACKAGE}
\title{errum: Bayesian Estimation of an Exploratory Reduced Reparameterized Unified Model}
\description{
Perform a bayesian estimation of the Exploratory Reduced
    Reparameterized Unified Model (errum) described by Culpepper and Chen (2018)
    <doi:10.3102/1076998618791306>
}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (0000-0003-4226-6176) [copyright holder]
  \item Jeffrey A. Douglas \email{jeffdoug@illinois.edu}
}

}
