% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errum-package.R
\docType{package}
\name{errum-package}
\alias{errum-package}
\alias{_PACKAGE}
\title{errum: Exploratory Reduced Reparameterized Unified Model Estimation}
\description{
Perform a Bayesian estimation of the exploratory reduced
    reparameterized unified model (ErRUM) described by Culpepper and Chen (2018)
    <doi:10.3102/1076998618791306>.
}
\seealso{
\code{\link[=errum]{errum()}}
}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID}) [copyright holder]
  \item Jeffrey A. Douglas \email{jeffdoug@illinois.edu}
}

}
\keyword{internal}
